Pendência estilização página detalhes
