+++ pokedex/src/Pages/DetailPage/Styled.js
@@ -1,12 +1,6 @@
-import React, { useState, useEffect } from 'react'
-import { useNavigate } from 'react-router-dom'
-import { goBack, goToHomePage } from '../routes/Coordinator'
-import styled from 'styled-components'
-import Imagem from '../img/Header2.png'
-import { useContext } from 'react'
-import { Pokemon } from '../GlobalState/Context'
+import styled from "styled-components"
 
-const HeaderContainer = styled.div`
+export const HeaderContainer = styled.div`
 display: flex;
 flex-direction: column;
 align-items: center;
@@ -15,9 +9,8 @@ img{
 width: 60vw;
 }
 `
-const SubHeader = styled.div`
+export const SubHeader = styled.div`
 display: flex;
-
 flex-direction: row;
 justify-content: space-around;
 button{
@@ -56,7 +49,6 @@ z-index: -1;
 background-color: #ffcb05!important;
 box-shadow: 0 -4px rgb(34 38 44 / 50%) inset, 0 4px rgb(34 38 44 / 99%) inset, -4px 0 rgb(34 38 44 / 50%) inset, 4px 0 rgb(34 38 44 / 50%) inset;
 }
-
 :after {
 position: absolute;
 pointer-events: none;
@@ -69,41 +61,33 @@ content: '';
 box-shadow: 0 4px 0 0 rgb(0 0 0 / 15%);
 transition: .7s cubic-bezier(0,.8,.26,.99);
 }
-
 :hover:before {
 box-shadow: 0 -4px rgb(34 38 44 / 50%) inset, 0 4px rgb(34 38 44 / 20%) inset, -4px 0 rgb(34 38 44 / 20%) inset, 4px 0 rgb(34 38 44 / 50%) inset;
 }
-
 :hover:after {
 box-shadow: 0 4px 0 0 rgb(0 0 0 / 15%);
-
 }
-
 :active {
 transform: translateY(4px);
-
 }
-
 :active:after {
 box-shadow: 0 0px 0 0 rgb(0 0 0 / 15%);
-
 }
-
 }
 `
 
-const DetailContainer = styled.div`
+export const DetailContainer = styled.div`
 color: white;
 display: flex;
 flex-direction: column;
-
 `
-const SubDetail = styled.div`
+
+export const SubDetail = styled.div`
 display: flex;
 flex-direction: row;
 justify-content: space-around;
 `
-const CardPokemon = styled.div`
+export const CardPokemon = styled.div`
 display: flex;
 width: 400px;
 height: 800px;
@@ -115,9 +99,9 @@ border-radius: 10px;
 img{
 width: 350px;
 }
-
 `
-const Habilidades = styled.div`
+
+export const Habilidades = styled.div`
 display: flex;
 width: 250px;
 justify-content: space-between;
@@ -127,10 +111,9 @@ border: 5px double #dedede ;
 border-radius: 10px;
 height: 100px;
 text-align: center;
-
-
 `
-const Tipos= styled.div`
+
+export const Tipos = styled.div`
 display: flex;
 width: 250px;
 justify-content: space-between;
@@ -140,7 +123,8 @@ border: 5px double #dedede ;
 border-radius: 10px;
 height: 125px;
 `
-const Movimentos = styled.div`
+
+export const Movimentos = styled.div`
 display: flex;
 width: 250px;
 justify-content: space-between;
@@ -148,64 +132,4 @@ flex-direction: column;
 align-items: center;
 border: 5px double #dedede ;
 border-radius: 10px;
-
 `
-
-
-export default function DetailsPage() {
-  const navigate = useNavigate()
-
-  const paramns = useContext(Pokemon)
-  return (
-    <div>
-      <HeaderContainer>
-        <img src={Imagem} />
-      </HeaderContainer>
-      <SubHeader>
-        <button onClick={() => goBack(navigate)}>Voltar</button>
-        <button onClick={() => goToHomePage(navigate)}>Home</button>
-      </SubHeader>
-      <DetailContainer>
-
-        {paramns.detailsPokemon.name &&
-          <SubDetail>
-            <CardPokemon>
-            <div>{paramns.detailsPokemon.name}</div>
-            <img src={paramns.detailsPokemon.sprites.other.dream_world.front_default} />
-            <img src={paramns.detailsPokemon.sprites.back_default} />
-            </CardPokemon>
-            <Habilidades>
-            <div>Habilidades: {paramns.detailsPokemon.abilities.map((habilidade) => {
-              return (
-                <p>{habilidade.ability.name}</p>
-                )
-              })}
-            </div>
-              </Habilidades>
-              <Tipos>
-            <div>
-              <p>Tipo:</p>
-              {paramns.detailsPokemon.types.map((tipos) => {
-                return (
-                  <p>{tipos.type.name}</p>
-                  )
-                })}
-            </div>
-                </Tipos>
-                <Movimentos>
-            <div>
-              <p>Movimentos:</p>
-              {paramns.detailsPokemon.moves.map((movimentos, index) => {
-                return (
-                  index < 4 &&
-                  <p>{movimentos.move.name}</p>
-                  )
-                })}
-            </div>
-                </Movimentos>
-          </SubDetail>
-        }
-      </DetailContainer>
-    </div>
-  )
-}
