+++ pokedex/src/Pages/HomePage/Styled.js
@@ -1,11 +1,6 @@
-import React, { useContext } from 'react'
-import { useNavigate } from 'react-router-dom'
-import { goToPokedexPage, goToDetailPage } from '../routes/Coordinator'
-import styled from 'styled-components'
-import Imagem from '../img/Header2.png'
-import { Pokemon } from '../GlobalState/Context'
+import styled from "styled-components"
 
-const HeaderContainer = styled.div`
+export const HeaderContainer = styled.div`
 display: flex;
 flex-direction: column;
 align-items: center;
@@ -15,7 +10,7 @@ width: 60vw;
 }
 `
 
-const SubHeader = styled.div`
+export const SubHeader = styled.div`
 display: flex;
 flex-direction: column;
 align-items: center;
@@ -55,7 +50,6 @@ z-index: -1;
 background-color: #ffcb05!important;
 box-shadow: 0 -4px rgb(34 38 44 / 50%) inset, 0 4px rgb(34 38 44 / 99%) inset, -4px 0 rgb(34 38 44 / 50%) inset, 4px 0 rgb(34 38 44 / 50%) inset;
 }
-
 :after {
 position: absolute;
 pointer-events: none;
@@ -68,30 +62,22 @@ content: '';
 box-shadow: 0 4px 0 0 rgb(0 0 0 / 15%);
 transition: .7s cubic-bezier(0,.8,.26,.99);
 }
-
 :hover:before {
 box-shadow: 0 -4px rgb(34 38 44 / 50%) inset, 0 4px rgb(34 38 44 / 20%) inset, -4px 0 rgb(34 38 44 / 20%) inset, 4px 0 rgb(34 38 44 / 50%) inset;
 }
-
 :hover:after {
 box-shadow: 0 4px 0 0 rgb(0 0 0 / 15%);
-
 }
-
 :active {
 transform: translateY(4px);
-
 }
-
 :active:after {
 box-shadow: 0 0px 0 0 rgb(0 0 0 / 15%);
-
 }
-
 }
 `
 
-const CardContainer = styled.div`
+export const CardContainer = styled.div`
 display: flex;
 color:white;
 list-style-type:none;
@@ -103,7 +89,7 @@ gap: 20px;
 justify-content: center;
 `
 
-const Cardzinhos = styled.div`
+export const Cardzinhos = styled.div`
 display: flex;
 width: 300px;
 height: 300px;
@@ -112,16 +98,15 @@ flex-direction: column;
 align-items: center;
 border: 5px double #dedede ;
 border-radius: 10px;
-
 border-bottom-right-radius:0;
 border-bottom-left-radius: 0;
 img{
 max-width: 180px;
 max-height: 180px;
 }
-
 `
-const Header2 = styled.div`
+
+export const Header2 = styled.div`
 background-color: #dc0a2d;
 width: 300px;
 border-radius: 7px;
@@ -132,8 +117,8 @@ text-align: center;
 font-size: 20px;
 }
 `
-const FooterCard = styled.div`
 
+export const FooterCard = styled.div`
 button{
 font-family: 'Press Start 2P', cursive;
 width: 150px;
@@ -171,7 +156,6 @@ z-index: -1;
 background-color: #dc0a2d!important;
 box-shadow: 0 -4px rgb(34 38 44 / 50%) inset, 0 4px rgb(34 38 44 / 99%) inset, -4px 0 rgb(34 38 44 / 50%) inset, 4px 0 rgb(34 38 44 / 50%) inset;
 }
-
 :after {
 position: absolute;
 pointer-events: none;
@@ -184,67 +168,17 @@ content: '';
 box-shadow: 0 4px 0 0 rgb(0 0 0 / 15%);
 transition: .7s cubic-bezier(0,.8,.26,.99);
 }
-
 :hover:before {
 box-shadow: 0 -4px rgb(0 0 0 / 50%) inset, 0 4px rgb(255 255 255 / 20%) inset, -4px 0 rgb(255 255 255 / 20%) inset, 4px 0 rgb(0 0 0 / 50%) inset;
 }
-
 :hover:after {
 box-shadow: 0 4px 0 0 rgb(0 0 0 / 15%);
-
 }
-
 :active {
 transform: translateY(4px);
-
 }
-
 :active:after {
 box-shadow: 0 0px 0 0 rgb(0 0 0 / 15%);
-
 }
-
 }
 `
-
-export default function HomePage() {
-  const navigate = useNavigate()
-
-  const paramns = useContext(Pokemon)
-
-  const twoFunction = (navigate, id) => {
-    paramns.getPokemonDetailFunction(id)
-    goToDetailPage(navigate)
-  }
-
-  const listPokemons = paramns.detail && paramns.detail.map((pokemon) => {
-    return (
-      <Cardzinhos key={pokemon.id}>
-        <Header2>
-          <h3>{pokemon.name.toUpperCase()}</h3>
-        </Header2>
-        <img src={pokemon.sprites.other.dream_world.front_default} />
-        <FooterCard>
-          <button onClick={() => paramns.addPokedex(pokemon.id)}>Adicionar</button>
-          <button onClick={() => twoFunction(navigate, pokemon.id)}>Detalhes</button>
-        </FooterCard>
-
-      </Cardzinhos>
-    )
-  })
-
-  return (
-    <div>
-      <HeaderContainer>
-        <img src={Imagem} />
-      </HeaderContainer>
-      <SubHeader>
-        <button onClick={() => goToPokedexPage(navigate)}>Pokédex</button>
-      </SubHeader>
-      <CardContainer>
-        {listPokemons}
-      </CardContainer>
-    </div>
-  )
-}
-
